Switch to GitHub actions (#544)
* Add ci_osx action

* Add ci_ubuntu action

* Remove Travis and AppVeyor
Both are now replaced with Github Actions. Although Travis and AppVeyor 
were working fine, the main reason to do this is because with Github 
Actions the continuous integration can be run without creating a Pull 
Request.

* Add plyer_deploy action

* Add GitHub Actions Badges to README

* Add ci_windows action

* Add ci_ubuntu.sh

* Add ci_osx.sh

* Add ci_windows.ps1

* Remove Style from ci_osx and ci_windows
Style is needed only once (in this case, in ci_ubuntu)

* Remove `twine` from `devrequirements`
`twine` is used in the `plyer_deploy` action and is installed using `deployment_dependencies`, located in `ci_ubuntu.sh`

* Add style_dependencies in ci_ubuntu.sh
This because `style` will run in a separate job, so just the needed dependencies are installed in there.

* Create a job for `style` only
Considering that `style` only needs to run using a single Python version and uses less dependencies, having it as a separate job should finish faster and let the user know quicker if any linter warning appears.

* Rename `build` job to `tests`

* Check style with flake8

* Create artifacts in plyer_deploy

* Create a draft release in plyer_deploy

* Add upload coverage step

* Build the package before creating the artifacts

* Generalize the Python version used in plyer_deploy

* Run tests job with Python 3.x instead of using matrix
In order to avoid the `Upload Coverage` step to run in three different Python version, the `tests` job now runs with Python 3.x, just like in Kivy. To keep consistency, this change is also applied in `ci_osx` and `ci_windows`.